Integration and unit tests in separate folders
